requirements.txt
Django>=5.0,<6.0
WeasyPrint>=61.0
gunicorn>=23.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
